+++ lib/add_route/widgets/takes_counter.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:redpoint/widgets/add_form/small_icon_button.dart';
+import 'package:redpoint/add_route/widgets/small_icon_button.dart';
 
 class TakesCounter extends StatefulWidget {
   const TakesCounter({super.key});
