--- lib/pages/profile_page.dart
@@ -1,6 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:redpoint/widgets/layout/page_template.dart';
-
-class ProfilePage extends PageTemplate {
-  ProfilePage() : super(title: "Profile", body: const Center(child: Text("Profile")));
-}
