--- lib/pages/social_page.dart
@@ -1,6 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:redpoint/widgets/layout/page_template.dart';
-
-class SocialPage extends PageTemplate {
-  SocialPage() : super(title: "Social", body: const Center(child: Text("Social")));
-}
