+++ lib/profile/profile_page.dart
@@ -1,0 +1,7 @@
+import 'package:flutter/material.dart';
+import 'package:redpoint/shared/widgets/layout/page_template.dart';
+
+class ProfilePage extends PageTemplate {
+  ProfilePage()
+      : super(title: "Profile", body: const Center(child: Text("Profile")));
+}
