+++ lib/shared/model/completed_status.dart
