+++ lib/shared/model/difficulty.dart
