+++ lib/shared/model/status.dart
