+++ lib/social/social_page.dart
@@ -1,0 +1,7 @@
+import 'package:flutter/material.dart';
+import 'package:redpoint/shared/widgets/layout/page_template.dart';
+
+class SocialPage extends PageTemplate {
+  SocialPage()
+      : super(title: "Social", body: const Center(child: Text("Social")));
+}
